PR 2.4: analysis collaborator migration
Move the entire Analysis class (~1006 LoC) from
src/pyramids/dataset/ops/analysis.py into
src/pyramids/dataset/_collaborators.py:Analysis. Migrated members:

* Public methods: stats, count_domain_cells, apply, fill, extract,
  overlay, get_mask, footprint, get_histogram, plot.
* Public staticmethod: normalize.
* Private helpers: _get_stats, _process_color_table, _rescale.

Inside the collaborator, every read of Dataset state goes through
self._ds.<state> via the weakref proxy. The substituted attributes
were: _check_alignment, _iloc, _update_inplace, band_count, band_names,
bbox, columns, create_from_array, crop, crs, epsg, gdal_dtype,
geotransform, get_band_by_color, logger, map_to_array_coordinates,
no_data_value, raster, read_array, read_overview_array, rows.

Three call-site fixups required:

* apply built the result with ``type(self)._build_dataset(...)`` and
  fill with ``type(self).dataset_like(...)``. Under the proxy,
  ``type(self._ds)`` returns ``weakref.ProxyType``; both now use
  ``self._ds.__class__.<X>(...)`` so the proxy's transparent attribute
  access dispatches to the wrapped Dataset class.
* stats forks on ``mask is not None``: the masked branch crops the
  dataset and calls ``dst._get_stats(...)`` on the result. After
  migration, that becomes ``dst.analysis._get_stats(...)`` (route the
  private helper through the collaborator).
* apply(inplace=True) and fill(inplace=True) on the original mixin
  returned ``self`` (the Dataset). The collaborator now returns
  ``None`` for those paths, and the Dataset.apply / Dataset.fill facades
  substitute ``self`` for the None return so callers still get
  identity preservation (``ds.apply(f, inplace=True) is ds`` works).

src/pyramids/dataset/dataset.py: drop _AnalysisMixin from import alias
list and from Dataset's MRO. Add eleven facades (stats,
count_domain_cells, apply, fill, extract, overlay, get_mask, footprint,
get_histogram, plot) thin-forwarding to self.analysis.<X>. The apply
and fill facades carry the None-substitution described above.

src/pyramids/dataset/ops/__init__.py: drop Analysis import / __all__
entry. src/pyramids/dataset/ops/analysis.py: keep module-level imports
only; the Analysis class is gone.

tests/:

* test_dataset_unit.py: import Analysis alongside Vectorize. Tests that
  reach for Dataset.normalize / Dataset._rescale / ds._get_stats now
  use Analysis.normalize / Analysis._rescale /
  ds.analysis._get_stats.
* test_collaborators.py: move 10 analysis rows from FORWARDING_METHODS
  to FACADE_METHODS.

Full main test suite passes; only the two pre-existing Windows
file-lock flakes in test_dataset_collection.py remain failing.
