+++ src/pyramids/dataset/ops/__init__.py
@@ -1,12 +1,10 @@
 """Dataset operation mixins."""
 
-from pyramids.dataset.ops.analysis import Analysis
 from pyramids.dataset.ops.band_metadata import BandMetadata
 from pyramids.dataset.ops.io import IO
 from pyramids.dataset.ops.spatial import Spatial
 
 __all__ = [
-    "Analysis",
     "BandMetadata",
     "IO",
     "Spatial",
